Add files via upload
This project is all about to operate a rescue team as quick as possible for our bravest ARMY Soldiers in war field or any panic situation. It is just a working model of our concept i.e. a wearable IoT device similar to a wristband which detects some reading and sends the data to the nearest basecamp, based on that the rescue team will be operated.